refactor: query policy 위치 정리
