+++ src/shared/api/index.ts
@@ -4,3 +4,4 @@ export * from "./execution";
 export * from "./oauth-token";
 export * from "./template";
 export * from "./workflow";
+export * from "./query-policy";
