+++ src/shared/model/workflow/useAddWorkflowNodeMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from "@tanstack/react-query";
 
 import type { NodeAddRequest } from "../../api";
 import { workflowApi } from "../../api";
-import { type MutationPolicyOptions, toMutationMeta } from "../query-policy";
+import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
 
 import { syncWorkflowCache } from "./workflow-cache-utils";
 
@@ -30,3 +30,4 @@ export const useAddWorkflowNodeMutation = (
       await options?.onError?.(error, variables, onMutateResult, context);
     },
   });
+
