+++ src/shared/model/workflow/useCreateWorkflowMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { workflowApi } from "../../api";
-import { type MutationPolicyOptions, toMutationMeta } from "../query-policy";
+import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
 
 import { syncWorkflowCache } from "./workflow-cache-utils";
 
@@ -23,3 +23,4 @@ export const useCreateWorkflowMutation = (
       await options?.onError?.(error, variables, onMutateResult, context);
     },
   });
+
