+++ src/shared/model/workflow/useDeleteWorkflowMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { workflowApi } from "../../api";
-import { type MutationPolicyOptions, toMutationMeta } from "../query-policy";
+import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
 
 import { removeWorkflowDomainCache } from "./workflow-cache-utils";
 
@@ -20,3 +20,4 @@ export const useDeleteWorkflowMutation = (
       await options?.onError?.(error, workflowId, onMutateResult, context);
     },
   });
+
