+++ src/shared/model/workflow/useUpdateWorkflowNodeMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from "@tanstack/react-query";
 
 import type { NodeUpdateRequest } from "../../api";
 import { workflowApi } from "../../api";
-import { type MutationPolicyOptions, toMutationMeta } from "../query-policy";
+import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
 
 import { syncWorkflowCache } from "./workflow-cache-utils";
 
@@ -31,3 +31,4 @@ export const useUpdateWorkflowNodeMutation = (
       await options?.onError?.(error, variables, onMutateResult, context);
     },
   });
+
